First complete parsing and insertion of all data in galaxy_json.
- Station IDs present in galaxy_json have been found in duplicate,
  generate internal mapping based on key being system_station.
- Guarantee that all stations have a matching economy with a default.
- Still missing data from transform_bodies, some debugging required.
- Update EDDB, station IDs are now big integers. All FKs updated.
- Remove large object to ID mapping files they can be generated.
- Delete old update database logic. Replaced with bulk_insert_mappings
  calls.
- During generall process, now an intermediary step gathering all
  unique factions was required.
- Gather faction names from controllingFaction field in stations, some
  factions are not present in "factions" field of a system but only
  mentioned in controllingFaction.
- Such straggler factions will be written to a separate file for later
  update.

+++ cogdb/eddb.py
@@ -554,7 +554,7 @@ class StationFeatures(ReprMixin, TimestampMixin, UpdatableMixin, Base):
         'repair', 'shipyard', 'technology_broker', 'universal_cartographics', 'updated_at'
     ]
 
-    id = sqla.Column(sqla.Integer, sqla.ForeignKey('stations.id'), primary_key=True)
+    id = sqla.Column(sqla.BigInteger, sqla.ForeignKey('stations.id'), primary_key=True)
 
     blackmarket = sqla.Column(sqla.Boolean)
     carrier_administration = sqla.Column(sqla.Boolean)
@@ -606,7 +606,7 @@ class StationEconomy(ReprMixin, Base):
     __tablename__ = "station_economies"
     _repr_keys = ['id', 'economy_id', 'primary', 'proportion']
 
-    id = sqla.Column(sqla.Integer, sqla.ForeignKey('stations.id'), primary_key=True)
+    id = sqla.Column(sqla.BigInteger, sqla.ForeignKey('stations.id'), primary_key=True)
     economy_id = sqla.Column(sqla.Integer, sqla.ForeignKey('economies.id'), primary_key=True)
     primary = sqla.Column(sqla.Boolean, primary_key=True, default=False)
     proportion = sqla.Column(sqla.Float)
@@ -627,7 +627,7 @@ class Station(ReprMixin, TimestampMixin, UpdatableMixin, Base):
         'type_id', 'system_id', 'controlling_minor_faction_id', 'updated_at'
     ]
 
-    id = sqla.Column(sqla.Integer, primary_key=True)
+    id = sqla.Column(sqla.BigInteger, primary_key=True)
     controlling_minor_faction_id = sqla.Column(sqla.Integer, sqla.ForeignKey('factions.id'))
     system_id = sqla.Column(sqla.Integer)
     type_id = sqla.Column(sqla.Integer, nullable=False)
@@ -908,9 +908,9 @@ class Conflict(ReprMixin, TimestampMixin, UpdatableMixin, Base):
     status_id = sqla.Column(sqla.Integer, sqla.ForeignKey('conflict_states.id'))
     type_id = sqla.Column(sqla.Integer, sqla.ForeignKey('conflict_states.id'))
     faction1_id = sqla.Column(sqla.Integer, sqla.ForeignKey('factions.id'), primary_key=True)
-    faction1_stake_id = sqla.Column(sqla.Integer, sqla.ForeignKey('stations.id'))
+    faction1_stake_id = sqla.Column(sqla.BigInteger, sqla.ForeignKey('stations.id'))
     faction2_id = sqla.Column(sqla.Integer, sqla.ForeignKey('factions.id'), primary_key=True)
-    faction2_stake_id = sqla.Column(sqla.Integer, sqla.ForeignKey('stations.id'))
+    faction2_stake_id = sqla.Column(sqla.BigInteger, sqla.ForeignKey('stations.id'))
 
     faction1_days = sqla.Column(sqla.Integer)
     faction2_days = sqla.Column(sqla.Integer)
